Conflicts:
	admin-dev/themes/default/css/admin.css
	admin-dev/themes/default/template/controllers/carrier_wizard/helpers/form/form.tpl
	admin-dev/themes/default/template/controllers/carrier_wizard/helpers/form/form_ranges.tpl
	admin-dev/themes/default/template/controllers/categories/helpers/form/form.tpl
	admin-dev/themes/default/template/controllers/home/content.tpl
	admin-dev/themes/default/template/controllers/localization/content.tpl
	admin-dev/themes/default/template/controllers/orders/_customized_data.tpl
	admin-dev/themes/default/template/controllers/orders/_new_product.tpl
	admin-dev/themes/default/template/controllers/orders/_product_line.tpl
	admin-dev/themes/default/template/controllers/orders/helpers/view/view.tpl
	admin-dev/themes/default/template/controllers/tax_rules/helpers/list/list_header.tpl
	admin-dev/themes/default/template/footer.tpl
	controllers/admin/AdminCarrierWizardController.php
	controllers/admin/AdminCarriersController.php
	controllers/admin/AdminCmsController.php
	controllers/admin/AdminLocalizationController.php
	controllers/admin/AdminManufacturersController.php
	controllers/admin/AdminSuppliersController.php
	css/admin.css
	js/admin-products.js
	js/admin_carrier_wizard.js
	translations/fr.gzip

+++ classes/Language.php
@@ -70,7 +70,7 @@ class LanguageCore extends ObjectModel
 	/** @var array Languages cache */
 	protected static $_checkedLangs;
 	protected static $_LANGUAGES;
-	protected static $countActiveLanguages;
+	protected static $countActiveLanguages = array();
 
 	protected	$webserviceParameters = array(
 		'objectNodeName' => 'language',
@@ -111,18 +111,22 @@ class LanguageCore extends ObjectModel
 		$iso_code = $newIso ? $newIso : $this->iso_code;
 
 		if (!file_exists(_PS_TRANSLATIONS_DIR_.$iso_code))
-			mkdir(_PS_TRANSLATIONS_DIR_.$iso_code);
+		{
+			if (@mkdir(_PS_TRANSLATIONS_DIR_.$iso_code))
+				@chmod(_PS_TRANSLATIONS_DIR_.$iso_code, 0777);
+		}
+
 		foreach ($this->translationsFilesAndVars as $file => $var)
 		{
 			$path_file = _PS_TRANSLATIONS_DIR_.$iso_code.'/'.$file.'.php';
 			if (!file_exists($path_file))
 				if ($file != 'tabs')
-					file_put_contents($path_file, '<?php
+					@file_put_contents($path_file, '<?php
 	global $'.$var.';
 	$'.$var.' = array();
 ?>');
 				else
-					file_put_contents($path_file, '<?php
+					@file_put_contents($path_file, '<?php
 	$'.$var.' = array();
 	return $'.$var.';
 ?>');
@@ -460,7 +464,8 @@ class LanguageCore extends ObjectModel
 				}
 			closedir($handle);
 		}
-		rmdir($dir);
+		if (is_writable($dir))
+			rmdir($dir);
 	}
 
 	public function delete()
@@ -519,7 +524,7 @@ class LanguageCore extends ObjectModel
 				'-default-'.ImageType::getFormatedName('medium').'.jpg',
 				'-default-'.ImageType::getFormatedName('small').'.jpg'
 			);
-			$image_directories = array(_PS_CAT_IMG_DIR_, _PS_MANU_IMG_DIR_, _PS_PROD_IMG_DIR_, _PS_SUPP_IMG_DIR_);
+			$images_directories = array(_PS_CAT_IMG_DIR_, _PS_MANU_IMG_DIR_, _PS_PROD_IMG_DIR_, _PS_SUPP_IMG_DIR_);
 			foreach ($images_directories as $image_directory)
 				foreach ($images as $image)
 				{
@@ -756,15 +761,18 @@ class LanguageCore extends ObjectModel
 		return (isset(self::$_cache_language_installation[$iso_code]) ? self::$_cache_language_installation[$iso_code] : false);
 	}
 
-	public static function countActiveLanguages()
+	public static function countActiveLanguages($id_shop = null)
 	{
-		if (!self::$countActiveLanguages)
-			self::$countActiveLanguages = Db::getInstance()->getValue('
+		if ($id_shop === null)
+			$id_shop = (int)Context::getContext()->shop->id;
+
+		if (!isset(self::$countActiveLanguages[$id_shop]))
+			self::$countActiveLanguages[$id_shop] = Db::getInstance()->getValue('
 				SELECT COUNT(DISTINCT l.id_lang) FROM `'._DB_PREFIX_.'lang` l
-				'.Shop::addSqlAssociation('lang', 'l').'
+				JOIN '._DB_PREFIX_.'lang_shop lang_shop ON (lang_shop.id_lang = l.id_lang AND lang_shop.id_shop = '.(int)$id_shop.')
 				WHERE l.`active` = 1
 			');
-		return self::$countActiveLanguages;
+		return self::$countActiveLanguages[$id_shop];
 	}
 
 	public static function downloadAndInstallLanguagePack($iso, $version = null, $params = null)
@@ -780,13 +788,22 @@ class LanguageCore extends ObjectModel
 		$lang_pack_ok = false;
 		$errors = array();
 		$file = _PS_TRANSLATIONS_DIR_.$iso.'.gzip';
+
 		if (!$lang_pack_link = Tools::file_get_contents('http://www.prestashop.com/download/lang_packs/get_language_pack.php?version='.$version.'&iso_lang='.Tools::strtolower($iso)))
 			$errors[] = Tools::displayError('Archive cannot be downloaded from prestashop.com.');
 		elseif (!$lang_pack = Tools::jsonDecode($lang_pack_link))
 			$errors[] = Tools::displayError('Error occurred when language was checked according to your Prestashop version.');
 		elseif ($content = Tools::file_get_contents('http://translations.prestashop.com/download/lang_packs/gzip/'.$lang_pack->version.'/'.Tools::strtolower($lang_pack->iso_code.'.gzip')))
 			if (!@file_put_contents($file, $content))
-				$errors[] = Tools::displayError('Server does not have permissions for writing.');
+			{
+				if (is_writable(dirname($file)))
+				{
+					@unlink($file);
+					@file_put_contents($file, $content);
+				}
+				elseif (!is_writable($file))
+					$errors[] = Tools::displayError('Server does not have permissions for writing.').' ('.$file.')';
+			}
 		if (file_exists($file))
 		{
 			$gz = new Archive_Tar($file, true);
@@ -819,8 +836,8 @@ class LanguageCore extends ObjectModel
 	 * @since 1.5.0
 	 * @return bool
 	 */
-	public static function isMultiLanguageActivated()
+	public static function isMultiLanguageActivated($id_shop = null)
 	{
-		return (Language::countActiveLanguages() > 1);
+		return (Language::countActiveLanguages($id_shop) > 1);
 	}
 }
